working on creating list and adding new products to list

+++ ProductCatalogue/Book.cs
@@ -11,16 +11,16 @@ namespace ProductCatalogue
 
         private string _title { get => Title; }
         private string _author { get => Author; }
-        private int _pageLength { get => PageLength; }
+        private double _pageLength { get => PageLength; }
         private string _synopsis { get => Synopsis; }
 
         public string Title;
         public string Author;
-        public int PageLength;
+        public double PageLength;
         public string Synopsis;
 
 
-        public Book(string title, string author, int pageLength, string synopsis, double price) : base(price)
+        public Book(string title, string author, double pageLength, string synopsis, double price) : base(price)
         {
             Title = title;
             Author = author;
@@ -29,29 +29,6 @@ namespace ProductCatalogue
 
 
         }
-
-
-        //public void createBook(string title, string author, int pageLength, string synopsis, double price)
-        //{
-
-        //    Console.WriteLine("What is the title of the book?");
-        //    string Title = Console.ReadLine();
-
-        //    Console.WriteLine("Who is the author?");
-        //    string Author = Console.ReadLine();
-
-        //    Console.WriteLine("How many pages does your book have?");
-        //    double PageLength = Convert.ToInt32(Console.ReadLine());
-
-        //    Console.WriteLine("What is the synopsis?");
-        //    string Synopsis = Console.ReadLine();
-
-        //    Console.WriteLine("What is the price");
-        //    double Price = Convert.ToInt32(Console.ReadLine());
-
-        //    Console.WriteLine($" {this.Title} by {this.Author}: {this.Synopsis}, {this.PageLength} pages, ${this.Price}");
-
-        //}
 
     }
 }
