+++ ProductCatalogue/Movie.cs
@@ -11,28 +11,21 @@ namespace ProductCatalogue
 
         private string _title { get => Title; }
         private string _director { get => Director; }
-        private int _runtime { get => Runtime; }
+        private double _runtime { get => Runtime; }
         private string _synposis { get => Synopsis; }
 
         public string Title;
         public string Director;
-        public int Runtime;
+        public double Runtime;
         public string Synopsis;
 
 
-        public Movie(string title, string director, int runtime, string synopsis, double price) : base(price)
+        public Movie(string title, string director, double runtime, string synopsis, double price) : base(price)
         {
             Title = title;
             Director = director;
            Runtime = runtime;
           Synopsis = synopsis;
-
-
-
-            //Console.WriteLine($" {title} by {director}: {synopsis}, {runtime}, ${price}");
-
-            //Catalogue._catalogueMain.Add(
-            // new Movie(newMovie, );
 
         }
     }
